function/requirements.txt
requests
pymysql
python-dotenv
